+++ common/steamcraft/common/core/helper/CommonHelper.java
@@ -46,13 +46,13 @@ public class CommonHelper {
 	    float f = 1.0F;
 	    float f1 = player.prevRotationPitch + (player.rotationPitch - player.prevRotationPitch) * f;
 	    float f2 = player.prevRotationYaw + (player.rotationYaw - player.prevRotationYaw) * f;
-	    double d0 = player.prevPosX + (player.posX - player.prevPosX) * (double) f;
-	    double d1 = player.prevPosY + (player.posY - player.prevPosY) * (double) f;
+	    double d0 = player.prevPosX + (player.posX - player.prevPosX) * f;
+	    double d1 = player.prevPosY + (player.posY - player.prevPosY) * f;
 	    
 	    if (!world.isRemote && player instanceof EntityPlayer)
 	        d1 += 1.62D;
 	    
-	    double d2 = player.prevPosZ + (player.posZ - player.prevPosZ) * (double) f;
+	    double d2 = player.prevPosZ + (player.posZ - player.prevPosZ) * f;
 	    Vec3 vec3 = world.getWorldVec3Pool().getVecFromPool(d0, d1, d2);
 	    float f3 = MathHelper.cos(-f2 * 0.017453292F - (float) Math.PI);
 	    float f4 = MathHelper.sin(-f2 * 0.017453292F - (float) Math.PI);
@@ -66,7 +66,7 @@ public class CommonHelper {
 	        d3 = ((EntityPlayerMP) player).theItemInWorldManager.getBlockReachDistance();
 	    }
 	    
-	    Vec3 vec31 = vec3.addVector((double) f7 * d3, (double) f6 * d3, (double) f8 * d3);
+	    Vec3 vec31 = vec3.addVector(f7 * d3, f6 * d3, f8 * d3);
 	    return world.rayTraceBlocks_do_do(vec3, vec31, par3, !par3);
 	}
 }
